refactor: Clean-up of DefaultDataQueryService [DHIS2-13644]
