+++ dhis-2/dhis-services/dhis-service-analytics/src/main/java/org/hisp/dhis/analytics/data/DefaultDataQueryService.java
@@ -27,18 +27,17 @@
  */
 package org.hisp.dhis.analytics.data;
 
+import static java.util.stream.Collectors.toList;
+import static org.hisp.dhis.analytics.AnalyticsAggregationType.fromAggregationType;
 import static org.hisp.dhis.analytics.DataQueryParams.DISPLAY_NAME_ATTRIBUTEOPTIONCOMBO;
 import static org.hisp.dhis.analytics.DataQueryParams.DISPLAY_NAME_CATEGORYOPTIONCOMBO;
-import static org.hisp.dhis.analytics.DataQueryParams.DISPLAY_NAME_DATA_X;
 import static org.hisp.dhis.analytics.DataQueryParams.DISPLAY_NAME_LATITUDE;
 import static org.hisp.dhis.analytics.DataQueryParams.DISPLAY_NAME_LONGITUDE;
-import static org.hisp.dhis.analytics.DataQueryParams.DISPLAY_NAME_ORGUNIT;
-import static org.hisp.dhis.analytics.DataQueryParams.DISPLAY_NAME_ORGUNIT_GROUP;
-import static org.hisp.dhis.analytics.DataQueryParams.DISPLAY_NAME_PERIOD;
-import static org.hisp.dhis.analytics.DataQueryParams.KEY_DE_GROUP;
-import static org.hisp.dhis.analytics.DataQueryParams.KEY_IN_GROUP;
 import static org.hisp.dhis.analytics.DataQueryParams.getMeasureCriteriaFromParam;
-import static org.hisp.dhis.analytics.util.AnalyticsUtils.throwIllegalQueryEx;
+import static org.hisp.dhis.analytics.OutputFormat.ANALYTICS;
+import static org.hisp.dhis.common.DimensionType.ATTRIBUTE_OPTION_COMBO;
+import static org.hisp.dhis.common.DimensionType.CATEGORY_OPTION_COMBO;
+import static org.hisp.dhis.common.DimensionType.STATIC;
 import static org.hisp.dhis.common.DimensionalObject.ATTRIBUTEOPTIONCOMBO_DIM_ID;
 import static org.hisp.dhis.common.DimensionalObject.CATEGORYOPTIONCOMBO_DIM_ID;
 import static org.hisp.dhis.common.DimensionalObject.DATA_X_DIM_ID;
@@ -47,49 +46,30 @@ import static org.hisp.dhis.common.DimensionalObject.LONGITUDE_DIM_ID;
 import static org.hisp.dhis.common.DimensionalObject.ORGUNIT_DIM_ID;
 import static org.hisp.dhis.common.DimensionalObject.ORGUNIT_GROUP_DIM_ID;
 import static org.hisp.dhis.common.DimensionalObject.PERIOD_DIM_ID;
-import static org.hisp.dhis.common.DimensionalObjectUtils.asList;
-import static org.hisp.dhis.common.DimensionalObjectUtils.asTypedList;
 import static org.hisp.dhis.common.DimensionalObjectUtils.getDimensionalItemIds;
-import static org.hisp.dhis.common.IdentifiableObjectUtils.getLocalPeriodIdentifier;
-import static org.hisp.dhis.commons.collection.ListUtils.sort;
-import static org.hisp.dhis.i18n.I18nFormat.FORMAT_DATE;
-import static org.hisp.dhis.organisationunit.OrganisationUnit.KEY_LEVEL;
-import static org.hisp.dhis.organisationunit.OrganisationUnit.KEY_ORGUNIT_GROUP;
-import static org.hisp.dhis.organisationunit.OrganisationUnit.KEY_USER_ORGUNIT;
-import static org.hisp.dhis.organisationunit.OrganisationUnit.KEY_USER_ORGUNIT_CHILDREN;
-import static org.hisp.dhis.organisationunit.OrganisationUnit.KEY_USER_ORGUNIT_GRANDCHILDREN;
-import static org.hisp.dhis.period.DailyPeriodType.ISO_FORMAT;
+import static org.hisp.dhis.common.DimensionalObjectUtils.getItemsFromParam;
+import static org.hisp.dhis.common.DisplayProperty.NAME;
+import static org.hisp.dhis.common.IdScheme.UID;
+import static org.hisp.dhis.feedback.ErrorCode.E7125;
+import static org.springframework.util.Assert.notNull;
 
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Date;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import lombok.RequiredArgsConstructor;
 
-import org.hisp.dhis.analytics.AnalyticsAggregationType;
-import org.hisp.dhis.analytics.AnalyticsFinancialYearStartKey;
 import org.hisp.dhis.analytics.AnalyticsSecurityManager;
 import org.hisp.dhis.analytics.DataQueryParams;
 import org.hisp.dhis.analytics.DataQueryService;
 import org.hisp.dhis.analytics.OrgUnitField;
-import org.hisp.dhis.analytics.OutputFormat;
-import org.hisp.dhis.calendar.Calendar;
 import org.hisp.dhis.category.CategoryOptionCombo;
 import org.hisp.dhis.common.AnalyticalObject;
 import org.hisp.dhis.common.BaseDimensionalObject;
-import org.hisp.dhis.common.BaseNameableObject;
-import org.hisp.dhis.common.CodeGenerator;
 import org.hisp.dhis.common.DataQueryRequest;
-import org.hisp.dhis.common.DimensionItemKeywords;
-import org.hisp.dhis.common.DimensionService;
-import org.hisp.dhis.common.DimensionType;
 import org.hisp.dhis.common.DimensionalItemObject;
 import org.hisp.dhis.common.DimensionalObject;
 import org.hisp.dhis.common.DimensionalObjectUtils;
@@ -97,33 +77,14 @@ import org.hisp.dhis.common.DisplayProperty;
 import org.hisp.dhis.common.EventDataQueryRequest;
 import org.hisp.dhis.common.IdScheme;
 import org.hisp.dhis.common.IdentifiableObjectManager;
-import org.hisp.dhis.common.IdentifiableProperty;
 import org.hisp.dhis.common.IllegalQueryException;
-import org.hisp.dhis.dataelement.DataElementGroup;
-import org.hisp.dhis.feedback.ErrorCode;
 import org.hisp.dhis.feedback.ErrorMessage;
-import org.hisp.dhis.hibernate.HibernateProxyUtils;
-import org.hisp.dhis.i18n.I18n;
 import org.hisp.dhis.i18n.I18nFormat;
 import org.hisp.dhis.i18n.I18nManager;
-import org.hisp.dhis.indicator.IndicatorGroup;
 import org.hisp.dhis.organisationunit.OrganisationUnit;
-import org.hisp.dhis.organisationunit.OrganisationUnitGroup;
-import org.hisp.dhis.organisationunit.OrganisationUnitService;
-import org.hisp.dhis.period.Period;
-import org.hisp.dhis.period.PeriodType;
-import org.hisp.dhis.period.RelativePeriodEnum;
-import org.hisp.dhis.period.RelativePeriods;
-import org.hisp.dhis.period.WeeklyPeriodType;
-import org.hisp.dhis.period.comparator.AscendingPeriodComparator;
-import org.hisp.dhis.security.acl.AclService;
-import org.hisp.dhis.setting.SettingKey;
-import org.hisp.dhis.setting.SystemSettingManager;
-import org.hisp.dhis.user.CurrentUserService;
 import org.hisp.dhis.user.User;
 import org.hisp.dhis.util.ObjectUtils;
 import org.springframework.stereotype.Service;
-import org.springframework.util.Assert;
 
 /**
  * @author Lars Helge Overland
@@ -133,23 +94,11 @@ import org.springframework.util.Assert;
 public class DefaultDataQueryService
     implements DataQueryService
 {
-    public static final Collection<DateTimeFormatter> DATE_TIME_FORMATTER = Stream.of( FORMAT_DATE, ISO_FORMAT )
-        .map( DateTimeFormatter::ofPattern )
-        .collect( Collectors.toList() );
+    private final DimensionalObjectCreator dimensionalObjectCreator;
 
     private final IdentifiableObjectManager idObjectManager;
 
-    private final OrganisationUnitService organisationUnitService;
-
-    private final DimensionService dimensionService;
-
     private final AnalyticsSecurityManager securityManager;
-
-    private final SystemSettingManager systemSettingManager;
-
-    private final AclService aclService;
-
-    private final CurrentUserService currentUserService;
 
     private final I18nManager i18nManager;
 
@@ -164,7 +113,7 @@ public class DefaultDataQueryService
 
         DataQueryParams.Builder params = DataQueryParams.newBuilder();
 
-        IdScheme inputIdScheme = ObjectUtils.firstNonNull( request.getInputIdScheme(), IdScheme.UID );
+        IdScheme inputIdScheme = ObjectUtils.firstNonNull( request.getInputIdScheme(), UID );
 
         if ( request.getDimension() != null && !request.getDimension().isEmpty() )
         {
@@ -192,7 +141,7 @@ public class DefaultDataQueryService
 
         if ( request.getAggregationType() != null )
         {
-            params.withAggregationType( AnalyticsAggregationType.fromAggregationType( request.getAggregationType() ) );
+            params.withAggregationType( fromAggregationType( request.getAggregationType() ) );
         }
 
         return params
@@ -216,7 +165,7 @@ public class DefaultDataQueryService
             .withOutputIdScheme( request.getOutputIdScheme() )
             .withOutputDataElementIdScheme( request.getOutputDataElementIdScheme() )
             .withOutputOrgUnitIdScheme( request.getOutputOrgUnitIdScheme() )
-            .withOutputFormat( OutputFormat.ANALYTICS )
+            .withOutputFormat( ANALYTICS )
             .withDuplicatesOnly( request.isDuplicatesOnly() )
             .withApprovalLevel( request.getApprovalLevel() )
             .withApiVersion( request.getApiVersion() )
@@ -227,12 +176,12 @@ public class DefaultDataQueryService
     @Override
     public DataQueryParams getFromAnalyticalObject( AnalyticalObject object )
     {
-        Assert.notNull( object, "Analytical object cannot be null" );
+        notNull( object, "Analytical object cannot be null" );
 
         DataQueryParams.Builder params = DataQueryParams.newBuilder();
 
         I18nFormat format = i18nManager.getI18nFormat();
-        IdScheme idScheme = IdScheme.UID;
+        IdScheme idScheme = UID;
         Date date = object.getRelativePeriodDate();
 
         String userOrgUnit = object.getRelativeOrganisationUnit() != null
@@ -272,7 +221,6 @@ public class DefaultDataQueryService
         Date relativePeriodDate, String userOrgUnit, I18nFormat format, IdScheme inputIdScheme )
     {
         List<DimensionalObject> list = new ArrayList<>();
-
         List<OrganisationUnit> userOrgUnits = getUserOrgUnits( null, userOrgUnit );
 
         if ( dimensionParams != null )
@@ -308,7 +256,7 @@ public class DefaultDataQueryService
     public DimensionalObject getDimension( String dimension, List<String> items, Date relativePeriodDate,
         List<OrganisationUnit> userOrgUnits, I18nFormat format, boolean allowNull, IdScheme inputIdScheme )
     {
-        return getDimension( dimension, items, relativePeriodDate, DisplayProperty.NAME, userOrgUnits,
+        return getDimension( dimension, items, relativePeriodDate, NAME, userOrgUnits,
             format, allowNull, inputIdScheme );
     }
 
@@ -316,333 +264,51 @@ public class DefaultDataQueryService
         DisplayProperty displayProperty, List<OrganisationUnit> userOrgUnits, I18nFormat format, boolean allowNull,
         IdScheme inputIdScheme )
     {
-        boolean allItems = items.isEmpty();
-        User user = currentUserService.getCurrentUser();
-
         if ( DATA_X_DIM_ID.equals( dimension ) )
         {
-            List<DimensionalItemObject> dataDimensionItems = new ArrayList<>();
-
-            DimensionItemKeywords dimensionalKeywords = new DimensionItemKeywords();
-
-            for ( String uid : items )
-            {
-                if ( uid.startsWith( KEY_DE_GROUP ) ) // DATA ELEMENT GROUP
-                {
-                    String groupUid = DimensionalObjectUtils.getUidFromGroupParam( uid );
-
-                    DataElementGroup group = idObjectManager.getObject( DataElementGroup.class, inputIdScheme,
-                        groupUid );
-
-                    if ( group != null )
-                    {
-                        dataDimensionItems.addAll( group.getMembers() );
-                        dimensionalKeywords.addKeyword( group );
-                    }
-                }
-                else if ( uid.startsWith( KEY_IN_GROUP ) ) // INDICATOR GROUP
-                {
-                    String groupUid = DimensionalObjectUtils.getUidFromGroupParam( uid );
-
-                    IndicatorGroup group = idObjectManager.getObject( IndicatorGroup.class, inputIdScheme, groupUid );
-
-                    if ( group != null )
-                    {
-                        dataDimensionItems.addAll( group.getMembers() );
-                        dimensionalKeywords.addKeyword( group );
-                    }
-                }
-                else
-                {
-                    DimensionalItemObject dimItemObject = dimensionService.getDataDimensionalItemObject(
-                        inputIdScheme, uid );
-
-                    if ( dimItemObject != null )
-                    {
-                        dataDimensionItems.add( dimItemObject );
-                    }
-                }
-            }
-
-            if ( dataDimensionItems.isEmpty() )
-            {
-                throwIllegalQueryEx( ErrorCode.E7124, DimensionalObject.DATA_X_DIM_ID );
-            }
-
-            return new BaseDimensionalObject( dimension, DimensionType.DATA_X, null, DISPLAY_NAME_DATA_X,
-                dataDimensionItems, dimensionalKeywords );
+            return dimensionalObjectCreator.fromDx( dimension, items, inputIdScheme );
         }
-
         else if ( CATEGORYOPTIONCOMBO_DIM_ID.equals( dimension ) )
         {
-            return new BaseDimensionalObject( dimension, DimensionType.CATEGORY_OPTION_COMBO, null,
+            return new BaseDimensionalObject( dimension, CATEGORY_OPTION_COMBO, null,
                 DISPLAY_NAME_CATEGORYOPTIONCOMBO, getCategoryOptionComboList( items, inputIdScheme ) );
         }
-
         else if ( ATTRIBUTEOPTIONCOMBO_DIM_ID.equals( dimension ) )
         {
-            return new BaseDimensionalObject( dimension, DimensionType.ATTRIBUTE_OPTION_COMBO, null,
+            return new BaseDimensionalObject( dimension, ATTRIBUTE_OPTION_COMBO, null,
                 DISPLAY_NAME_ATTRIBUTEOPTIONCOMBO, getCategoryOptionComboList( items, inputIdScheme ) );
         }
-
         else if ( PERIOD_DIM_ID.equals( dimension ) )
         {
-            Calendar calendar = PeriodType.getCalendar();
-            I18n i18n = i18nManager.getI18n();
-            List<Period> periods = new ArrayList<>();
-
-            DimensionItemKeywords dimensionalKeywords = new DimensionItemKeywords();
-
-            AnalyticsFinancialYearStartKey financialYearStart = systemSettingManager
-                .getSystemSetting( SettingKey.ANALYTICS_FINANCIAL_YEAR_START, AnalyticsFinancialYearStartKey.class );
-
-            boolean containsRelativePeriods = false;
-
-            for ( String isoPeriod : items )
-            {
-                // Contains isoPeriod and timeField
-                IsoPeriodHolder isoPeriodHolder = IsoPeriodHolder.of( isoPeriod );
-
-                if ( RelativePeriodEnum.contains( isoPeriodHolder.getIsoPeriod() ) )
-                {
-                    containsRelativePeriods = true;
-                    RelativePeriodEnum relativePeriod = RelativePeriodEnum.valueOf( isoPeriodHolder.getIsoPeriod() );
-
-                    dimensionalKeywords.addKeyword( isoPeriodHolder.getIsoPeriod(),
-                        i18n.getString( isoPeriodHolder.getIsoPeriod() ) );
-
-                    List<Period> relativePeriods = RelativePeriods.getRelativePeriodsFromEnum( relativePeriod,
-                        relativePeriodDate, format, true, financialYearStart );
-
-                    // If custom time filter is specified, set it in periods
-                    if ( isoPeriodHolder.hasDateField() )
-                    {
-                        relativePeriods.forEach( period -> period.setDateField( isoPeriodHolder.getDateField() ) );
-                    }
-
-                    periods.addAll( relativePeriods );
-                }
-                else
-                {
-                    Period period = PeriodType.getPeriodFromIsoString( isoPeriodHolder.getIsoPeriod() );
-
-                    if ( period != null )
-                    {
-                        if ( isoPeriodHolder.hasDateField() )
-                        {
-                            period.setDescription( isoPeriodHolder.getIsoPeriod() );
-                            period.setDateField( isoPeriodHolder.getDateField() );
-                        }
-
-                        dimensionalKeywords.addKeyword( isoPeriodHolder.getIsoPeriod(),
-                            format != null ? i18n.getString( format.formatPeriod( period ) )
-                                : isoPeriodHolder.getIsoPeriod() );
-
-                        periods.add( period );
-                    }
-                    else
-                    {
-                        Optional<Period> optionalPeriod = isoPeriodHolder.toDailyPeriod();
-                        if ( optionalPeriod.isPresent() )
-                        {
-                            Period periodToAdd = optionalPeriod.get();
-                            String startDate = i18nManager.getI18nFormat().formatDate( periodToAdd.getStartDate() );
-                            String endDate = i18nManager.getI18nFormat().formatDate( periodToAdd.getEndDate() );
-                            dimensionalKeywords.addKeyword(
-                                isoPeriodHolder.getIsoPeriod(),
-                                String.join( " - ", startDate, endDate ) );
-                            periods.add( periodToAdd );
-                        }
-                    }
-                }
-            }
-
-            // Remove duplicates
-            periods = periods.stream().distinct().collect( Collectors.toList() );
-
-            if ( containsRelativePeriods )
-            {
-                periods.sort( new AscendingPeriodComparator() );
-            }
-
-            for ( Period period : periods )
-            {
-                String name = format != null ? format.formatPeriod( period ) : null;
-
-                if ( !period.getPeriodType().getName().contains( WeeklyPeriodType.NAME ) )
-                {
-                    period.setShortName( name );
-                }
-
-                period.setName( name );
-
-                if ( !calendar.isIso8601() )
-                {
-                    period.setUid( getLocalPeriodIdentifier( period, calendar ) );
-                }
-            }
-
-            return new BaseDimensionalObject( dimension, DimensionType.PERIOD, null, DISPLAY_NAME_PERIOD,
-                asList( periods ), dimensionalKeywords );
+            return dimensionalObjectCreator.fromPeriod( dimension, items, relativePeriodDate, format );
         }
-
         else if ( ORGUNIT_DIM_ID.equals( dimension ) )
         {
-            List<DimensionalItemObject> ous = new ArrayList<>();
-            List<Integer> levels = new ArrayList<>();
-            List<OrganisationUnitGroup> groups = new ArrayList<>();
-
-            for ( String ou : items )
-            {
-                if ( KEY_USER_ORGUNIT.equals( ou ) && userOrgUnits != null && !userOrgUnits.isEmpty() )
-                {
-                    ous.addAll( userOrgUnits );
-                }
-                else if ( KEY_USER_ORGUNIT_CHILDREN.equals( ou ) && userOrgUnits != null && !userOrgUnits.isEmpty() )
-                {
-                    ous.addAll( OrganisationUnit.getSortedChildren( userOrgUnits ) );
-                }
-                else if ( KEY_USER_ORGUNIT_GRANDCHILDREN.equals( ou ) && userOrgUnits != null
-                    && !userOrgUnits.isEmpty() )
-                {
-                    ous.addAll( OrganisationUnit.getSortedGrandChildren( userOrgUnits ) );
-                }
-                else if ( ou != null && ou.startsWith( KEY_LEVEL ) )
-                {
-                    String level = DimensionalObjectUtils.getValueFromKeywordParam( ou );
-
-                    Integer orgUnitLevel = organisationUnitService.getOrganisationUnitLevelByLevelOrUid( level );
-
-                    if ( orgUnitLevel != null )
-                    {
-                        levels.add( orgUnitLevel );
-                    }
-                }
-                else if ( ou != null && ou.startsWith( KEY_ORGUNIT_GROUP ) )
-                {
-                    String uid = DimensionalObjectUtils.getUidFromGroupParam( ou );
-
-                    OrganisationUnitGroup group = idObjectManager.getObject( OrganisationUnitGroup.class, inputIdScheme,
-                        uid );
-
-                    if ( group != null )
-                    {
-                        groups.add( group );
-                    }
-                }
-                else if ( !inputIdScheme.is( IdentifiableProperty.UID ) || CodeGenerator.isValidUid( ou ) )
-                {
-                    OrganisationUnit unit = idObjectManager.getObject( OrganisationUnit.class, inputIdScheme, ou );
-
-                    if ( unit != null )
-                    {
-                        ous.add( unit );
-                    }
-                }
-            }
-
-            // Remove duplicates
-            ous = ous.stream().distinct().collect( Collectors.toList() );
-
-            List<DimensionalItemObject> orgUnits = new ArrayList<>();
-            List<OrganisationUnit> ousList = asTypedList( ous );
-            DimensionItemKeywords dimensionalKeywords = new DimensionItemKeywords();
-
-            if ( !levels.isEmpty() )
-            {
-                orgUnits.addAll( sort( organisationUnitService.getOrganisationUnitsAtLevels( levels, ousList ) ) );
-
-                dimensionalKeywords.addKeywords( levels.stream()
-                    .map( organisationUnitService::getOrganisationUnitLevelByLevel )
-                    .filter( Objects::nonNull )
-                    .collect( Collectors.toList() ) );
-            }
-
-            if ( !groups.isEmpty() )
-            {
-                orgUnits.addAll( sort( organisationUnitService.getOrganisationUnits( groups, ousList ) ) );
-
-                dimensionalKeywords.addKeywords( groups.stream()
-                    .map( group -> new BaseNameableObject( group.getUid(), group.getCode(),
-                        group.getDisplayProperty( displayProperty ) ) )
-                    .collect( Collectors.toList() ) );
-            }
-
-            // -----------------------------------------------------------------
-            // When levels / groups are present, OUs are considered boundaries
-            // -----------------------------------------------------------------
-
-            if ( levels.isEmpty() && groups.isEmpty() )
-            {
-                orgUnits.addAll( ous );
-            }
-
-            if ( !dimensionalKeywords.isEmpty() )
-            {
-                dimensionalKeywords.addKeywords( ousList );
-            }
-
-            if ( orgUnits.isEmpty() )
-            {
-                throwIllegalQueryEx( ErrorCode.E7124, DimensionalObject.ORGUNIT_DIM_ID );
-            }
-
-            // Remove duplicates
-            orgUnits = orgUnits.stream().distinct().collect( Collectors.toList() );
-
-            return new BaseDimensionalObject( dimension, DimensionType.ORGANISATION_UNIT,
-                null, DISPLAY_NAME_ORGUNIT, orgUnits, dimensionalKeywords );
+            return dimensionalObjectCreator.fromOrgUnit( dimension, items, displayProperty, userOrgUnits,
+                inputIdScheme );
         }
-
         else if ( ORGUNIT_GROUP_DIM_ID.equals( dimension ) )
         {
-            List<DimensionalItemObject> ougs = new ArrayList<>();
-
-            for ( String uid : items )
-            {
-                OrganisationUnitGroup organisationUnitGroup = idObjectManager.getObject( OrganisationUnitGroup.class,
-                    inputIdScheme, uid );
-
-                if ( organisationUnitGroup != null )
-                {
-                    ougs.add( organisationUnitGroup );
-                }
-            }
-
-            return new BaseDimensionalObject( dimension, DimensionType.ORGANISATION_UNIT_GROUP, null,
-                DISPLAY_NAME_ORGUNIT_GROUP, ougs );
+            return dimensionalObjectCreator.fromOrgUnitGroup( dimension, items, inputIdScheme );
         }
-
         else if ( LONGITUDE_DIM_ID.contains( dimension ) )
         {
-            return new BaseDimensionalObject( dimension, DimensionType.STATIC, null, DISPLAY_NAME_LONGITUDE,
+            return new BaseDimensionalObject( dimension, STATIC, null, DISPLAY_NAME_LONGITUDE,
                 new ArrayList<>() );
         }
-
         else if ( LATITUDE_DIM_ID.contains( dimension ) )
         {
-            return new BaseDimensionalObject( dimension, DimensionType.STATIC, null, DISPLAY_NAME_LATITUDE,
+            return new BaseDimensionalObject( dimension, STATIC, null, DISPLAY_NAME_LATITUDE,
                 new ArrayList<>() );
         }
-
         else
         {
-            DimensionalObject dimObject = idObjectManager.get( DataQueryParams.DYNAMIC_DIM_CLASSES, inputIdScheme,
-                dimension );
+            Optional<BaseDimensionalObject> baseDimensionalObject = dimensionalObjectCreator.fromDynamic( dimension,
+                items, displayProperty, inputIdScheme );
 
-            if ( dimObject != null && dimObject.isDataDimension() )
+            if ( baseDimensionalObject.isPresent() )
             {
-                Class<?> dimClass = HibernateProxyUtils.getRealClass( dimObject );
-
-                Class<? extends DimensionalItemObject> itemClass = DimensionalObject.DIMENSION_CLASS_ITEM_CLASS_MAP
-                    .get( dimClass );
-
-                List<DimensionalItemObject> dimItems = !allItems
-                    ? asList( idObjectManager.getOrdered( itemClass, inputIdScheme, items ) )
-                    : getCanReadItems( user, dimObject );
-
-                return new BaseDimensionalObject( dimObject.getDimension(), dimObject.getDimensionType(), null,
-                    dimObject.getDisplayProperty( displayProperty ), dimItems, allItems );
+                return baseDimensionalObject.get();
             }
         }
 
@@ -651,7 +317,7 @@ public class DefaultDataQueryService
             return null;
         }
 
-        throw new IllegalQueryException( new ErrorMessage( ErrorCode.E7125, dimension ) );
+        throw new IllegalQueryException( new ErrorMessage( E7125, dimension ) );
     }
 
     @Override
@@ -663,31 +329,30 @@ public class DefaultDataQueryService
 
         if ( userOrgUnit != null )
         {
-            units.addAll( DimensionalObjectUtils.getItemsFromParam( userOrgUnit ).stream()
+            units.addAll( getItemsFromParam( userOrgUnit ).stream()
                 .map( ou -> idObjectManager.get( OrganisationUnit.class, ou ) )
                 .filter( Objects::nonNull )
-                .collect( Collectors.toList() ) );
+                .collect( toList() ) );
         }
         else if ( currentUser != null && params != null && params.getUserOrgUnitType() != null )
         {
             switch ( params.getUserOrgUnitType() )
             {
             case DATA_CAPTURE:
-                units.addAll( currentUser.getOrganisationUnits().stream().sorted().collect( Collectors.toList() ) );
+                units.addAll( currentUser.getOrganisationUnits().stream().sorted().collect( toList() ) );
                 break;
             case DATA_OUTPUT:
                 units.addAll(
-                    currentUser.getDataViewOrganisationUnits().stream().sorted().collect( Collectors.toList() ) );
+                    currentUser.getDataViewOrganisationUnits().stream().sorted().collect( toList() ) );
                 break;
             case TEI_SEARCH:
-                units.addAll(
-                    currentUser.getTeiSearchOrganisationUnits().stream().sorted().collect( Collectors.toList() ) );
+                units.addAll( currentUser.getTeiSearchOrganisationUnits().stream().sorted().collect( toList() ) );
                 break;
             }
         }
         else if ( currentUser != null )
         {
-            units.addAll( currentUser.getOrganisationUnits().stream().sorted().collect( Collectors.toList() ) );
+            units.addAll( currentUser.getOrganisationUnits().stream().sorted().collect( toList() ) );
         }
 
         return units;
@@ -710,20 +375,6 @@ public class DefaultDataQueryService
         return items.stream()
             .map( item -> idObjectManager.getObject( CategoryOptionCombo.class, inputIdScheme, item ) )
             .filter( Objects::nonNull )
-            .collect( Collectors.toList() );
-    }
-
-    /**
-     * Returns only objects for which the user has data or metadata read access.
-     *
-     * @param user the user.
-     * @param object the {@link DimensionalObject}.
-     * @return a list of {@link DimensionalItemObject}.
-     */
-    private List<DimensionalItemObject> getCanReadItems( User user, DimensionalObject object )
-    {
-        return object.getItems().stream()
-            .filter( o -> aclService.canDataOrMetadataRead( user, o ) )
-            .collect( Collectors.toList() );
+            .collect( toList() );
     }
 }
